fix: [UF-335] 리프레시 삭제

+++ src/main/java/com/example/ufo_fi/v2/auth/application/oauth/CustomOAuth2UserService.java
@@ -56,7 +56,9 @@ public class CustomOAuth2UserService extends DefaultOAuth2UserService {
      */
     private OAuth2User login(User existUser) {
         Refresh refresh = existUser.getRefresh();
-        refreshManager.delete(refresh);
+        if(refresh != null){
+            refreshManager.delete(refresh);
+        }
 
         return CustomOAuth2User.from(existUser);
     }
